Move dm-verity logic into own crate
