+++ sextant/src/lib.rs
@@ -12,4 +12,5 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
+pub mod dm_verity;
 pub mod npk;
